Initial writeup to document the protocol implemented
WIP writeup describing the protocol I've observed and implemented
in aquaweb.py, so that others can implement similar things without
having to pore over the code.  But in case of discrepencies, the
code will be the final say since it's been working flawlessly for
years now in my own setup.